Change the way install.sh handles initial setup
This commit introduces the following changes:
* the cript requires user's confirmation to perform any changes
* user is no longer prompted specifically for confirmation when
the scrip overrides an existing .dotfiles symlink
* a new abort function is added and it is called when user does not want
the script to continue or when an existing .dotfile directory (not a symlink)
is detected in the user's home directory.

+++ install.sh
@@ -2,24 +2,27 @@
 
 dotfiles="$HOME/.dotfiles"
 current_dir=$(pwd)
-abort="Aborting installation"
 
-if [ -d $dotfiles ]; then
-    if [ -L $dotfiles ]; then
-        while true; do
-            echo "A link $dotfiles already exists"
-            read -p "Do you want to override it?" yn
-            case $yn in
-                [Yy]* ) break;;
-                [Nn]* ) echo $abort; exit;;
-                * ) echo "Please answer yes or no.";;
-            esac
-        done
-    else
-        echo "$dotfiles already exists, and it's a directory. $abort"
-        exit
-    fi
+abort() {
+    echo "Aborting installation"
+    exit
+}
+
+
+echo "This script may overwrite files in your home directory."
+read -p "Are you sure you want to continue? (Y/n)" -n 1 -r
+echo ""
+if ! [[ $REPLY =~ ^[Yy]$ ]]
+then
+    abort
 fi
+
+
+if [ -d $dotfiles ] && ! [ -L $dotfiles ]; then
+    echo "$dotfiles already exists, and it's a directory."
+    abort
+fi
+
 
 echo "Creating a symbolic link from $current_dir to $dotfiles"
 ln -Tsf $(pwd) $dotfiles
